# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Signage API Documentation",
        "description": "Tài liệu API dành cho dự án Digital Signage",
        "contact": {
            "email": "admin@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register-device": {
            "post": {
                "tags": [
                    "REGISTER-DEVICE"
                ],
                "summary": "API Đăng ký Thiết bị mới đầy đủ thông tin (Từ Mobile/Box)",
                "description": "Android Box gọi API này gửi thông tin lên để nằm trong danh sách chờ duyệt. Khi Admin duyệt trên Filament, gọi lại API này sẽ trả về Token.",
                "operationId": "6caaa82eae107156379c6f8fcdbd43f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_code",
                                    "name"
                                ],
                                "properties": {
                                    "device_code": {
                                        "description": "Mã phần cứng duy nhất (Mobile tự lấy ngầm)",
                                        "type": "string",
                                        "example": "BOX_8899_PRO"
                                    },
                                    "name": {
                                        "description": "Tên gợi nhớ (Thợ lắp máy gõ vào ô Input)",
                                        "type": "string",
                                        "example": "Màn hình LED Tầng 1"
                                    },
                                    "os_version": {
                                        "description": "Phiên bản hệ điều hành của Box (Mobile tự lấy ngầm)",
                                        "type": "string",
                                        "example": "Android 11"
                                    },
                                    "app_version": {
                                        "description": "Phiên bản của ứng dụng chạy trên Box (Mobile tự lấy ngầm)",
                                        "type": "string",
                                        "example": "v1.0.0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Đăng ký thành công, đang chờ phê duyệt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thiết bị đã gửi thông tin thành công, vui lòng chờ phê duyệt."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Đã phê duyệt - Trả Token về cho Box tự động lưu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        },
                                        "device_token": {
                                            "type": "string",
                                            "example": "chuoi_token_ngau_nhien_64_ky_tu..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/server-time": {
            "get": {
                "tags": [
                    "Box API"
                ],
                "summary": "API lấy thời gian hiện tại của Server",
                "description": "Cung cấp mốc thời gian chuẩn của Server để Android Box đồng bộ đồng hồ, tránh lỗi lệch múi giờ.",
                "operationId": "f34502fe043410f9be4a5fa73697e428",
                "responses": {
                    "200": {
                        "description": "Lấy thời gian thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "timezone": {
                                            "type": "string",
                                            "example": "Asia/Ho_Chi_Minh"
                                        },
                                        "server_time": {
                                            "type": "string",
                                            "example": "2026-06-15 03:15:00"
                                        },
                                        "timestamp": {
                                            "type": "integer",
                                            "example": 1781512500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deviceToken": []
                    }
                ]
            }
        },
        "/api/check-status": {
            "post": {
                "tags": [
                    "Box API"
                ],
                "summary": "API Check trạng thái Box (Heartbeat)",
                "description": "Cục Box gọi API này định kỳ để báo danh và cập nhật thời gian hoạt động gần nhất (last_connected_at) lên hệ thống.",
                "operationId": "49e788ea688fb2e4f9148de88d19ecab",
                "responses": {
                    "200": {
                        "description": "Cập nhật trạng thái thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cập nhật trạng thái thiết bị thành công."
                                        },
                                        "device_name": {
                                            "type": "string",
                                            "example": "Màn hình LED Sảnh Chính"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực hoặc Token sai"
                    }
                },
                "security": [
                    {
                        "deviceToken": []
                    }
                ]
            }
        },
        "/api/get-schedule": {
            "get": {
                "tags": [
                    "Box API"
                ],
                "summary": "Lấy lịch trình cho thiết bị qua Device Token",
                "operationId": "849cf2645b37362c21dabb7d1de91636",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "401": {
                        "description": "Token không hợp lệ hoặc thiếu"
                    }
                },
                "security": [
                    {
                        "deviceToken": []
                    }
                ]
            }
        },
        "/api/updateMediaLog": {
            "post": {
                "tags": [
                    "Box API"
                ],
                "summary": "Đồng bộ lịch sử phát (Logs) theo chùm Schedule",
                "operationId": "0cb5c626c705b5f7c28bee968062f20e",
                "requestBody": {
                    "description": "Dữ liệu danh sách log đã gom nhóm gọn gàng theo Schedule ID",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "logs"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "description": "ID của lịch trình đang phát",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "logs": {
                                        "description": "Danh sách các file media đã phát xong",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "media_id",
                                                "played_at"
                                            ],
                                            "properties": {
                                                "media_id": {
                                                    "description": "ID của file media",
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "played_at": {
                                                    "description": "Thời gian phát xong (Định dạng chuẩn: Y-m-d H:i:s)",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-06-23 17:01:20"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đồng bộ thành công dữ liệu sạch",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Hệ thống đã đồng bộ thành công 2 log hợp lệ!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Sai cấu trúc JSON hoặc sai định dạng thời gian"
                    },
                    "401": {
                        "description": "Token không hợp lệ hoặc thiếu token"
                    },
                    "400": {
                        "description": "Không có dòng log nào hợp lệ để lưu"
                    }
                },
                "security": [
                    {
                        "deviceToken": []
                    }
                ]
            }
        },
        "/api/info": {
            "get": {
                "tags": [
                    "Box API"
                ],
                "summary": "API Lấy thông tin cấu hình hệ thống của Box",
                "description": "Box gọi API này khi vừa khởi động để lấy thông tin định danh và tham số vận hành từ Server.",
                "operationId": "cd549b6a701f5c2f52763fea511a8d5c",
                "responses": {
                    "200": {
                        "description": "Lấy thông tin thành công"
                    }
                },
                "security": [
                    {
                        "deviceToken": []
                    }
                ]
            }
        },
        "/api/download-media/{id}": {
            "get": {
                "tags": [
                    "Box API"
                ],
                "summary": "API Stream/Tải tệp tin media",
                "description": "Hỗ trợ trả về tệp tin vật lý dưới dạng luồng bytes giúp thiết bị tua và tải video mượt mà.",
                "operationId": "a2d8b8cc1fe5359fbe79334b5e818149",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của file media",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trả về luồng dữ liệu file"
                    }
                },
                "security": [
                    {
                        "deviceToken": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "REGISTER-DEVICE",
            "description": "REGISTER-DEVICE"
        },
        {
            "name": "Box API",
            "description": "Box API"
        }
    ],
    "components": {
        "securitySchemes": {
            "deviceToken": {
                "type": "http",
                "scheme": "bearer",
                "description": "Nhập chuỗi Device Token (64 ký tự) của Android Box vào đây"
            }
        }
    }
}